--- a/base/base.jl
+++ b/base/base.jl
@@ -31,6 +31,7 @@ call(T::Type{NewvarNode}, s::Symbol) = Core.call(T, s)
 call(T::Type{TopNode}, s::Symbol) = Core.call(T, s)
 call(T::Type{Module}, args...) = Core.call(T, args...)
 call(T::Type{Task}, f::ANY) = Core.call(T, f)
+call(T::Type{GenSym}, n::Int) = Core.call(T, n)
 
 call{T}(::Type{T}, args...) = convert(T, args...)::T