--- a/base/boot.jl
+++ b/base/boot.jl
@@ -274,6 +274,7 @@ _new(:GotoNode, :Int)
 _new(:TopNode, :Symbol)
 _new(:NewvarNode, :Symbol)
 _new(:QuoteNode, :ANY)
+_new(:GenSym, :Int)
 
 Module(name::Symbol) = ccall(:jl_f_new_module, Any, (Any,), name)::Module
 Module() = Module(:anonymous)