this introduces a new Symbol type as a compiler optimization for places where only a simple unnamed memory location is required. it is for use in places where the jl_varinfo_t is known by construction:
```
     int closureidx = -1
     bool isAssigned = 1
     bool isCaptured = 0
     bool isSA = 1
     bool isVolatile = 0
     bool isArgument = 0
     bool isGhost = 0/1
     bool hasGCRoot = 0/1
     bool escapes = 0
     bool usedUndef = 0
     bool used = 0/1
```

the vinfo for the Symbol should correspond to `18`

TODO:
- [x] after expanding macros, run the jlgensym renumbering pass before doing code splicing
- [x] use `(make-jlgensym)` instead of `(gensym)` in julia-syntax generation passes
- [x] build a boxed value cache for GenSym

@JeffBezanson