@vtjnash 2015-01-24
(travis failures appears to be an unrelated issues with homebrew)

it's rare i can make a PR with such a clear across-the-board performance boost:

after:
```
jameson@julia:~/julia/test/perf$ make kernel
cons                 96.266  188.395  146.604   39.080
randn_zig            37.648   37.902   37.751    0.113
gk                  158.208  404.366  248.603   93.413
sparsemul            64.683   66.159   65.245    0.590
sparsemul2           98.068  160.006  112.229   26.794
sparserange          67.659   97.540   82.707   13.952
matvec               11.285   11.452   11.351    0.066
sortperm             15.733   15.777   15.758    0.016
stockcorr          6188.298 6223.274 6206.063   13.933
bench_eu_vec        145.298  191.906  155.934   20.140
bench_eu_devec       85.300   86.161   85.519    0.363
actorgraph          894.421  962.800  924.511   29.213
laplace_vec        1864.612 1941.088 1909.872   36.892
laplace_devec       162.415  164.409  163.389    0.795
go_benchmark       1704.725 1729.694 1722.933   10.303
simplex              56.799   57.757   57.386    0.353
raytracer          2549.374 2601.338 2569.424   26.118
funarg              173.549  227.696  195.820   28.767
vectorize           254.572  297.089  266.823   17.276
splitline            75.648  117.342  108.291   18.256
json                 32.894   78.704   42.742   20.109
add1                 91.500  127.591  119.742   15.796
devec_add1           13.797   14.003   13.889    0.096
add1_logical         85.611  126.161  116.746   17.494
devec_add1_logical   16.508   16.776   16.598    0.104

fib                   0.097    0.109    0.099    0.005
parse_int             0.359    0.398    0.377    0.015
mandel                0.341    0.342    0.341    0.000
quicksort             0.716    0.804    0.736    0.038
pi_sum               55.044   55.148   55.081    0.051
rand_mat_stat        25.393   64.662   41.392   18.443
rand_mat_mul         50.424   57.288   52.698    2.648
printfd              46.103   46.433   46.228    0.145
```

before (release 0.3.5):
```
cons                119.709  210.150  173.628   33.290
randn_zig            38.245   38.271   38.259    0.013
gk                  158.591  377.580  249.697   80.515
sparsemul            65.569   67.968   66.444    1.097
sparsemul2          104.697  165.051  118.196   26.232
sparserange          68.486  103.821   84.073   15.150
matvec               10.221   11.171   10.808    0.487
sortperm             18.779   18.816   18.804    0.015
stockcorr          6270.317 6315.420 6296.297   19.654
bench_eu_vec        193.549  232.460  212.060   17.300
bench_eu_devec       89.238   90.507   89.575    0.527
actorgraph         1162.488 1711.965 1488.358  243.940
laplace_vec        2688.123 2797.466 2741.339   39.592
laplace_devec       186.556  188.872  187.950    0.883
go_benchmark       1941.020 1953.543 1948.660    5.191
simplex              59.600   59.704   59.654    0.042
raytracer          2757.393 3121.404 2894.569  156.090
funarg              214.107  281.481  241.587   35.847
vectorize           318.323  364.866  333.030   19.164
splitline            93.737  147.382  133.132   22.837
json                 33.577   81.313   43.211   21.299
add1                 99.710  151.870  138.738   22.017
devec_add1           13.622   13.758   13.701    0.055
add1_logical        105.581  155.085  143.671   21.386
devec_add1_logical   19.555   19.594   19.567    0.016

fib                   0.134    0.135    0.135    0.000
parse_int             0.398    0.659    0.500    0.103
mandel                0.334    0.353    0.338    0.008
quicksort             0.695    0.720    0.709    0.012
pi_sum               55.027   55.033   55.031    0.002
rand_mat_stat        25.236   55.845   37.485   15.779
rand_mat_mul         51.349   56.827   53.778    2.059
printfd              46.784   46.892   46.825    0.042
```