@IainNZ 2015-01-24
Here are the percentages (new times / old times), using the data in the comments above:

```
                NAME     MIN     MAX    MEAN
                cons   84.70  101.44   90.28
           randn_zig   99.62   99.63   99.66
                  gk   97.17  103.73   97.08
           sparsemul   97.66   91.41   95.67
          sparsemul2   92.54   96.43   93.21
         sparserange  106.31  111.98  108.35
              matvec  104.19   99.43  101.41
            sortperm   82.49   80.28   81.15
           stockcorr  953.27  904.32  922.60
        bench_eu_vec   96.64  105.07   98.74
      bench_eu_devec   97.42   97.68   97.48
          actorgraph  101.74  104.45  103.12
         laplace_vec   97.37  100.13   99.07
       laplace_devec   88.35   85.47   86.96
        go_benchmark   98.18   98.12   98.64
             simplex   97.03   98.19   97.81
           raytracer  100.82  100.42  100.61
              funarg  100.12   99.70   99.72
           vectorize   96.01   98.99   97.80
           splitline   99.23   98.20   98.91
                json   98.63  107.06  103.01
                add1  106.57   97.61   99.15
          devec_add1   98.95   98.21   98.52
        add1_logical   92.53   87.29   92.09
  devec_add1_logical   84.84   86.14   85.25
                 fib   72.39   81.34   73.88
           parse_int   91.35   66.22   75.55
              mandel  102.40   96.07  100.89
           quicksort  101.85  111.20  103.37
              pi_sum   99.99   99.93   99.96
       rand_mat_stat  101.30  115.80  110.38
        rand_mat_mul   98.11  106.45  100.48
             printfd   98.69   97.62   98.45
```