@vtjnash 2015-01-24
> Whats up with stockcorr?

good catch, i had missed that one. turns out it was a fix on base that i hadn't incorporated into my test results

> What's the deal with the x.ji (large binary file) that's part of this commit?

hm, how did that sneak back in? I rebased to get rid of it, and added a gitignore entry for it. (it's the same as sys.ji, I just like to make copies of it)