@JeffBezanson 2015-01-24
Wow, is all that improvement just from having fewer symbol objects??