@JeffBezanson 2015-01-24
Quite nice to drop 2.5MB of just symbols from the system image. Yikes.