@vtjnash 2015-01-24
to shrink the image a bit further, we might also try to intern strings more often. for example, we end up with a lot of cases of "unrecognized keyword argument \"". and every LineNode seems to include a full copy of the filename string.