@vtjnash 2015-01-24
CI is green, so I'll go ahead and merge as soon as you are OK with it