@JeffBezanson 2015-01-24
I thought the line nodes used symbols for filenames.