@vtjnash 2015-01-24
> I thought the line nodes used symbols for filenames.

i thought so too, but the text seems to be showing up too many times in `strings sys.ji`