@JeffBezanson 2015-01-24
How do you feel about using the GenSym approach for all local variables? (In which case I would name it something like `LocalVar`.) This would have the following advantages:

1. Every local has an index, so we can use arrays with fast lookup for all types and variable properties. We can get rid of the current clunky varinfo arrays.
1. All local names become just a matter of debug info. This would automatically fix the problem we currently have of not reporting renamed variable names correctly in error messages.
1. Type info during inference becomes just a matrix.
1. Greater uniformity. Some people have complained, pretty reasonably, that when you see a symbol in an AST it is too hard to tell what it is.

Of course, we lose the advantage that `GenSym` means SSA. However it would be quite easy/fast to check `isSSA(vinfo[x.id])`.