@vtjnash 2015-01-24
it'll make an expaneded AST much harder to read, although I can see the appeal