@StefanKarpinski 2015-01-24
> it'll make an expaneded AST much harder to read, although I can see the appeal

Couldn't that be dealt with by printing the AST in the more human-readable form – i.e. with readable names replaced for gensyms where there is a name to be replaced?