@vtjnash 2015-01-25
unfortunately, one of the torture test i was hoping would be greatly improved by this PR (https://gist.github.com/vtjnash/27ff622b2cbed22b51dc) seems to have been not affected much / in the wrong direction:

after:
```
julia> @time include("../slow_performance.jl")
elapsed time: 0.024892235 seconds (333 kB allocated)
elapsed time: 3.61572336 seconds (7 MB allocated, 1.15% gc time in 1 pauses with 0 full sweep)

julia> @time show(t)
elapsed time: 24.334687017 seconds (2050 MB allocated, 10.09% gc time in 26 pauses with 13 full sweep)
2-element Array{T1Type,1}:
 
 

julia> 
```

before:
```
julia> @time include("../slow_performance.jl")
elapsed time: 0.023578148 seconds (334 kB allocated)
elapsed time: 3.447290547 seconds (7 MB allocated, 1.16% gc time in 1 pauses with 0 full sweep)

julia> @time show(t)
elapsed time: 50.349472299 seconds (2153 MB allocated, 5.71% gc time in 31 pauses with 13 full sweep)
2-element Array{T1Type,1}:
 
 

julia>
```